# docs/api/openapi.yaml
components:
  schemas:
...
    AnalysisHistoryResponse:
      description: Response model for analysis history of a drawing.
      properties:
        analyses:
          items:
            $ref: '#/components/schemas/AnomalyAnalysisResponse'
          title: Analyses
          type: array
        drawing_id:
          title: Drawing Id
          type: integer
        total_count:
          title: Total Count
          type: integer
      required:
      - drawing_id
      - analyses
      - total_count
      title: AnalysisHistoryResponse
      type: object
    AnalysisMethod:
      description: Enumeration for anomaly detection methods.
      enum:
      - autoencoder
      title: AnalysisMethod
      type: string
...
    AnomalyAnalysisResponse:
      description: Response model for anomaly analysis results.
      properties:
        age_group:
          description: Age group used for analysis
          title: Age Group
          type: string
        analysis_timestamp:
          format: date-time
          title: Analysis Timestamp
          type: string
        anomaly_score:
          description: Raw anomaly score
          title: Anomaly Score
          type: number
        confidence:
          description: Confidence in the anomaly decision
          maximum: 1.0
          minimum: 0.0
          title: Confidence
          type: number
        drawing_id:
          title: Drawing Id
          type: integer
        id:
          title: Id
          type: integer
        is_anomaly:
          description: Whether the drawing is flagged as anomalous
          title: Is Anomaly
          type: boolean
        method_used:
          $ref: '#/components/schemas/AnalysisMethod'
          description: Analysis method used
        normalized_score:
          description: 'Normalized anomaly score (0-100 scale: 0=no anomaly, 100=maximal
            anomaly)'
          maximum: 100.0
          minimum: 0.0
          title: Normalized Score
          type: number
        vision_model:
          $ref: '#/components/schemas/VisionModel'
          description: Vision model used
      required:
      - id
      - drawing_id
      - anomaly_score
      - normalized_score
      - is_anomaly
      - confidence
      - age_group
      - method_used
      - vision_model
      - analysis_timestamp
      title: AnomalyAnalysisResponse
      type: object
...
    Body_batch_upload_drawings_api_v1_drawings_batch_upload_post:
      properties:
        files:
          description: Multiple drawing files
          items:
            format: binary
            type: string
          title: Files
          type: array
      required:
      - files
      title: Body_batch_upload_drawings_api_v1_drawings_batch_upload_post
      type: object
...
    Body_upload_drawing_api_v1_drawings_upload_post:
      properties:
        age_years:
          description: Child's age in years
          maximum: 18.0
          minimum: 2.0
          title: Age Years
          type: number
        drawing_tool:
          anyOf:
          - type: string
          - type: 'null'
          description: Drawing tool used
          title: Drawing Tool
        expert_label:
          anyOf:
          - type: string
          - type: 'null'
          description: Expert assessment
          title: Expert Label
        file:
          description: Drawing image file (PNG, JPEG, BMP)
          format: binary
          title: File
          type: string
        prompt:
          anyOf:
          - type: string
          - type: 'null'
          description: Drawing prompt
          title: Prompt
        subject:
          anyOf:
          - type: string
          - type: 'null'
          description: Drawing subject
          title: Subject
      required:
      - file
      - age_years
      title: Body_upload_drawing_api_v1_drawings_upload_post
      type: object
...
    DrawingListResponse:
      description: Response model for listing multiple drawings.
      properties:
        drawings:
          items:
            $ref: '#/components/schemas/DrawingResponse'
          title: Drawings
          type: array
        page:
          description: Current page number
          minimum: 1.0
          title: Page
          type: integer
        page_size:
          description: Number of items per page
          maximum: 100.0
          minimum: 1.0
          title: Page Size
          type: integer
        total_count:
          title: Total Count
          type: integer
        total_pages:
          title: Total Pages
          type: integer
      required:
      - drawings
      - total_count
      - page
      - page_size
      - total_pages
      title: DrawingListResponse
      type: object
    DrawingResponse:
      description: Response model for drawing information.
      properties:
        age_years:
          title: Age Years
          type: number
        drawing_tool:
          anyOf:
          - type: string
          - type: 'null'
          title: Drawing Tool
        expert_label:
          anyOf:
          - type: string
          - type: 'null'
          title: Expert Label
        filename:
          title: Filename
          type: string
        id:
          title: Id
          type: integer
        prompt:
          anyOf:
          - type: string
          - type: 'null'
          title: Prompt
        subject:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
        upload_timestamp:
          format: date-time
          title: Upload Timestamp
          type: string
      required:
      - id
      - filename
      - age_years
      - subject
      - expert_label
      - drawing_tool
      - prompt
      - upload_timestamp
      title: DrawingResponse
      type: object
    ExpertLabel:
      description: Enumeration for expert labels on drawings.
      enum:
      - normal
      - concern
      - severe
      title: ExpertLabel
      type: string
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
    VisionModel:
      description: Enumeration for vision models.
      enum:
      - vit
      title: VisionModel
      type: string
info:
  contact:
    email: support@example.com
    name: API Support
  description: Machine learning system for detecting anomalies in children's drawings
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Children's Drawing Anomaly Detection System
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/v1/analysis/drawing/{drawing_id}:
    get:
      description: 'Get all analyses for a specific drawing.


        This endpoint returns the analysis history for a drawing,

        ordered by most recent first.'
      operationId: get_drawing_analyses_api_v1_analysis_drawing__drawing_id__get
      parameters:
      - in: path
        name: drawing_id
        required: true
        schema:
          title: Drawing Id
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 10
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/AnalysisHistoryResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Drawing Analyses
      tags:
      - analysis
...
  /api/v1/drawings/:
    get:
      description: List drawings with optional filtering and pagination.
      operationId: list_drawings_api_v1_drawings__get
      parameters:
      - in: query
        name: age_min
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Age Min
      - in: query
        name: age_max
        required: false
        schema:
          anyOf:
          - type: number
          - type: 'null'
          title: Age Max
      - in: query
        name: subject
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Subject
      - in: query
        name: expert_label
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ExpertLabel'
          - type: 'null'
          title: Expert Label
      - in: query
        name: page
        required: false
        schema:
          default: 1
          title: Page
          type: integer
      - in: query
        name: page_size
        required: false
        schema:
          default: 20
          title: Page Size
          type: integer
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/DrawingListResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List Drawings
      tags:
      - drawings
  /api/v1/drawings/batch/upload:
    post:
      description: 'Upload multiple drawings in batch.


        This endpoint accepts multiple files and processes them in the background.

        Returns an upload ID for tracking progress.'
      operationId: batch_upload_drawings_api_v1_drawings_batch_upload_post
      requestBody:
        content:
          multipart/form-data:
            example: {}
            schema:
              $ref: '#/components/schemas/Body_batch_upload_drawings_api_v1_drawings_batch_upload_post'
        required: true
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Batch Upload Drawings
      tags:
      - drawings
...
  /api/v1/drawings/upload:
    post:
      description: 'Upload drawing with metadata.


        This endpoint accepts multipart form data with an image file and metadata.

        The image is validated, preprocessed, and stored along with the metadata.'
      operationId: upload_drawing_api_v1_drawings_upload_post
      requestBody:
        content:
          multipart/form-data:
            example: {}
            schema:
              $ref: '#/components/schemas/Body_upload_drawing_api_v1_drawings_upload_post'
        required: true
      responses:
        '201':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/DrawingResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Upload Drawing
      tags:
      - drawings
  /api/v1/drawings/upload/progress/{upload_id}:
    get:
      description: Get upload progress for large file uploads.
      operationId: get_upload_progress_api_v1_drawings_upload_progress__upload_id__get
      parameters:
      - in: path
        name: upload_id
        required: true
        schema:
          title: Upload Id
          type: string
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Upload Progress
      tags:
      - drawings
  /api/v1/drawings/{drawing_id}:
    delete:
      description: Delete drawing and associated data.
      operationId: delete_drawing_api_v1_drawings__drawing_id__delete
      parameters:
      - in: path
        name: drawing_id
        required: true
        schema:
          title: Drawing Id
          type: integer
      responses:
        '204':
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Delete Drawing
      tags:
      - drawings
    get:
      description: Retrieve drawing details by ID.
      operationId: get_drawing_api_v1_drawings__drawing_id__get
      parameters:
      - in: path
        name: drawing_id
        required: true
        schema:
          title: Drawing Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/DrawingResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Drawing
      tags:
      - drawings
  /api/v1/drawings/{drawing_id}/file:
    get:
      description: Retrieve the actual drawing file.
      operationId: get_drawing_file_api_v1_drawings__drawing_id__file_get
      parameters:
      - in: path
        name: drawing_id
        required: true
        schema:
          title: Drawing Id
          type: integer
      responses:
        '200':
          content:
            application/json:
              example: {}
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              example: {}
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Get Drawing File
      tags:
      - drawings
...
servers:
- description: Development server
  url: http://localhost:8000
- description: API v1 base path
  url: /api/v1